+++ internal/fluentbit/config/builder/env_var_test.go
@@ -1,4 +1,4 @@
-package config
+package builder
 
 import (
 	"testing"
